// Website/EhsInput/src/models/MonthlyHealth.tsx
export interface MonthlyHealth {
    labourHrs: number;
    overtimeHrs: number;
    workingDays: number;
    hrsPerDay: number;
    fullTimeEqvTarifEmp: number;
    fullTimeEqvNonTarifEmp: number;
    sickDays: number;
    sickNotifications: number;
    longTermSick: number;
    plannedAbsentism: number;
    leaving: number;
}