// Website/EhsInput/src/models/MonthlySafety.tsx
export interface MonthlySafety {
    accidentsWithFirstAid: number;
    accidentsWithMedicalTreatment: number;
    accidentsWithLostTime: number;
    lostDays: number;
    nearMisses: number;
    safetyTrainings: number;
    safetyAudits: number;
    unsafeConditions: number;
    unsafeActs: number;
    fatalities: number;
    seriousInjuries: number;
    fullowUps: number;
}