// Website/EhsInput/src/models/MonthlyEnergy.tsx
export interface MonthlyEnergy {
    naturalGas: number;
    lightFuelOil: number;
    coal: number;
    carGasoline: number;
    carDiesel: number;
    bioFuel: number;
    electricityPurchased: number;
    sharePurchasedRenewable: number;
    selfGeneratedElectricityCosumed: number;
    selfGeneratedElectricitySold: number;
}