// Website/EhsInput/src/models/MonthlyWaste.tsx
export interface MonthlyWaste {
    fabricationWaste: number;
    productionWaste: number;
    otherNonHazardousWaste: number;
    reusedNonHazardousWaste: number;
    recycledNonHazardousWaste: number;
    hazardousWaste: number;
    reusedHazardousWaste: number;
    recycledHazardousWaste: number;
}